+++ NEWS.md
@@ -1,5 +1,6 @@
 # BrailleR 0.33.2
 - updated batch file creation so that they search for current version of R. Batch files should not require updating as a consequence.
+- start adding support for quarto.
 
 
 # BrailleR 0.33.1